+++ hardpy/pytest_hardpy/utils/dialog_box.py
@@ -9,10 +9,11 @@ from typing import Any
 class DialogBoxWidgetType(Enum):
     """Dialog box widget type."""
 
-    RADIOBUTTON = "radiobutton"
-    CHECKBOX = "checkbox"
     TEXT_INPUT = "textinput"
     NUMERIC_INPUT = "numericinput"
+    RADIOBUTTON = "radiobutton"
+    CHECKBOX = "checkbox"
+    IMAGE = "image"
 
 
 @dataclass
@@ -70,9 +71,7 @@ def generate_dialog_box_dict(dialog_box_data: DialogBox) -> dict:
     return data_dict
 
 
-def get_dialog_box_data(
-    input_data: str, widget: DialogBoxWidget | None
-) -> Any:
+def get_dialog_box_data(input_data: str, widget: DialogBoxWidget | None) -> Any:
     """Get the dialog box data in the correct format.
 
     Args:
@@ -96,5 +95,15 @@ def get_dialog_box_data(
                 return None
         case DialogBoxWidgetType.TEXT_INPUT:
             return input_data
+        case DialogBoxWidgetType.RADIOBUTTON:
+            try:
+                return input_data
+            except ValueError:
+                return None
+        case DialogBoxWidgetType.CHECKBOX:
+            try:
+                return input_data
+            except ValueError:
+                return None
         case _:
             return None